correct deleting features + create profile page and handle profile update

+++ src/pages/BoardPage.tsx
@@ -3,7 +3,7 @@ import { BoardType } from "@/utils/types";
 import { useNavigate, useParams } from "react-router-dom";
 import { ListsSection } from "../components/List/ListsSection";
 import { BoardPageSkeleton } from "@/components/skeletons";
-import { useDeleteDoc, useUpdateDoc } from "@/firebase/mutateHook";
+import { useDeleteBoard, useUpdateDoc } from "@/firebase/mutateHook";
 import { DeleteItem } from "@/components/Form/DeleteItem";
 import { UpdateTitle } from "@/components/Form/UpdateTitle";
 import { AddMember } from "@/components/Members/AddMember";
@@ -17,17 +17,16 @@ export const BoardPage = () => {
 
   
   const { data: board, isFetched } = useGetDoc<BoardType>("boards", boardId);
-  const deleteBoard = useDeleteDoc("boards", "boards", boardId)
-  const updateBoard = useUpdateDoc<BoardType>("board", "boards", boardId ?? "");
+  const deleteBoard = useDeleteBoard(currentUser?.id, boardId,);
+  const updateBoard = useUpdateDoc<BoardType>("board", "boards", boardId, undefined, ["boards", currentUser?.id ?? ""] );
 
   if (!boardId || !isFetched) {
     return <BoardPageSkeleton />;
   }
 
   const handleDelete = async () => {
-    deleteBoard.mutate([], {
-      onSuccess: () => naviagate("/"),
-    });
+    await deleteBoard.mutateAsync();
+    naviagate("/")
   };
 
   return (
@@ -42,14 +41,16 @@ export const BoardPage = () => {
               headingLevel={"h1"}
             />
             <div className="flex items-start gap-3 h-10">
+              <div className="h-10 flex items-center">
               <MembersDetails members={board.members} creator={board.creator} query={updateBoard} isBoard />
+              </div>
               {board.creator === currentUser?.id && <AddMember queryName="boards" queryId={board.creator} query={updateBoard} members={board.members}/>}
-              <DeleteItem
+              {board.creator === currentUser?.id && <DeleteItem
                 name="board"
                 handleDelete={handleDelete}
                 isText={false}
                 isPending={deleteBoard.isPending}
-              />
+              />}
             </div>
           </header>
           <ListsSection boardId={boardId} />
